moved group ranking loading from component to screen
